+++ libs/footer/src/lib/footer.tsx
@@ -81,7 +81,7 @@ export function Footer({
                       </Button>
                     )}
                   </li>
-                )
+                ),
               )}
             </StyledList>
           )}