+++ libs/circle-link/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog Circle Link component
 
+## [0.2.1](https://github.com/sauldeleon/portfolio-blog/compare/circle-link-0.2.0...circle-link-0.2.1) (2023-09-14)
+
 # [0.2.0](https://github.com/sauldeleon/portfolio-blog/compare/circle-link-0.1.2...circle-link-0.2.0) (2023-09-10)
 
 ### Bug Fixes