+++ libs/use-container-dimensions/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog useContainerDimensions hook
 
+## [0.1.1](https://github.com/sauldeleon/portfolio-blog/compare/use-container-dimensions-0.1.0...use-container-dimensions-0.1.1) (2023-09-14)
+
 # 0.1.0 (2023-09-02)
 
 ### Bug Fixes